testing removing inputs

+++ training/ddp_training/modules_ddp.py
@@ -500,20 +500,12 @@ class OGBoardEncoder(nn.Module):
         # Embeddings
         embeddings = torch.cat(
             [
-                self.turn_embeddings(turns),
-                self.white_kingside_castling_rights_embeddings(
-                    white_kingside_castling_rights
-                ),
-                self.white_queenside_castling_rights_embeddings(
-                    white_queenside_castling_rights
-                ),
-                self.black_kingside_castling_rights_embeddings(
-                    black_kingside_castling_rights
-                ),
-                self.black_queenside_castling_rights_embeddings(
-                    black_queenside_castling_rights
-                ),
-                self.board_position_embeddings(board_positions),
+                self.turn_embeddings(turns.to(torch.int64)).to(torch.float32),  # Ensure embeddings are float
+                self.white_kingside_castling_rights_embeddings(white_kingside_castling_rights.to(torch.int64)).to(torch.float32),
+                self.white_queenside_castling_rights_embeddings(white_queenside_castling_rights.to(torch.int64)).to(torch.float32),
+                self.black_kingside_castling_rights_embeddings(black_kingside_castling_rights.to(torch.int64)).to(torch.float32),
+                self.black_queenside_castling_rights_embeddings(black_queenside_castling_rights.to(torch.int64)).to(torch.float32),
+                self.board_position_embeddings(board_positions.to(torch.int64)).to(torch.float32),  
             ],
             dim=1,
         )  # (N, BOARD_STATUS_LENGTH, d_model)
